requirements.txt
aioredis==2.0.1
annotated-types==0.7.0
anyio==4.10.0
async-timeout==5.0.1
bidict==0.23.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
redis==6.4.0
requests==2.32.5
simple-websocket==1.1.0
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
wsproto==1.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
